TensorFlow2/Recommendation/DLRM_and_DCNv2/requirements.txt
git+https://github.com/NVIDIA/dllogger#egg=dllogger
absl-py>=0.7.0
pyarrow
pandas
joblib
tqdm
pyyaml
onnxruntime
git+https://github.com/onnx/tensorflow-onnx
tritonclient[all]==2.31
numpy<1.24
tabulate>=0.8.7
natsort>=7.0.0